Remove pointless 'recopying' of CFF (g)subr strings
For some reason, the Type 2 font serialisation (used for reassembling font
streams to pass to freetype) was using the FAPI callback to read the
(g)subr string data into it's main buffer, but then called the writing
function to write data to the buffer, using its own buffer as input -
effectively overwriting the bytes with the same bytes.

The only purpose (it seems) was to move the current stream position forward by
the appropriate amount.

Change it so we simply move the "current position" pointer on by the correct
offset.