Bug 702810: fix fapi buffer size parameter types
The buffer size parameters for the fapi callbacks used to be unsigned short,
but were changed to ints a while back. The calls for the (g)subr data in
Type 2/CFF fonts hadn't had the casts to ushort removed, causing the values
to be trucated.

There was also a Postscript/FAPI internal function that had not had the
buffer length parameter changed to int.