pdfwrite - fix memory leak with annotations
Annotations are somewhat odd and don't quite fit the scheme of other
cos objects. Each annotation is created as a dictionary, which is
stored in the Annots array (also a cos object)) of the page.

But we write the annots out to the file immediately, and free the
dictionary contents at that time. We cannot, however, free the dict
cos object, because we need the object number to write the Annots
array in the page dictionary, which we only do when writing the page
dictionary.

So at the end of the page the annots array contains a number of empty
dictionaries. We free the array, which would free the dictionary
contents (except there are none) but it does *not* free the dictionaries
because they have an ID......

So when we free the Annots array, we first need to set the ID of each
of the dictionary objects it contains to 0, so that the dictionary
will also be freed.