Fix error in token parsing in pdfwrite
I am trying to use a hex string <XXXXXXXX> in pdfi.
The code had a bug in it (probably because the ps code never generates
a hex string?) that was incrementing the token to be 1 longer than it
should have been.  Logic was just mangled.

I also simplified it to to take out the goto, which was egregious in
this case.  Just put in two lines of code where it belongs, instead of
a goto to those two lines of code...  (really?)