Work around for (I think) a VS2019 optimiser bug
VS2019 release builds crash with the input file from bug 702916 and several
other files, in copied_glyph_slot() because the pointer retrieved and stored
in *pslot is non-sensical.

Debug/Profile builds and optimised builds with earler VS versions don't show
the problem.

Adding debug code to assign the calculated index to an interim unsigned integer
variable also cause the problem to go away.

So, use that as a workaround.