Fix icc profile reference counting in transparency compositor
Found during Windows testing for a release.

The full test file for
https://bugs.ghostscript.com/show_bug.cgi?id=693365

would cause Ghostscript to crash due to an ICC profile being freed whilst a
reference was still being held for it. That was not counting up a reference
count when restoring the device profile back to a previous value.

Fixing that introduced a leak for other profiles. And that turned out to be
not decrementing the reference count when replacing a device profile.