Bug 702948: Ensure initial bitmap cache sizes are properly aligned
Now that the allocations from the bitmap cache are aligned to the platform's
required alignment, see commit:

260c443bc14cdffa4d94e14c3a57e35bebee3a5b

We also want the initial size of the memory pool used by the cache to be
"aligned".

This is so that code that attempts to identify cache entries to evict by
requesting a size equal to the entire size of cache memory pool doesn't get an
unexpected failure, because we've rounded up that allocation request to a value
larger than the entire size of the memory pool.

Because we don't expect an error to be possible at that point, a crash can
occur.

Of the "normal" platforms we use, this only exhibits on Win32 because that is
the only platform where the align_bitmap_mod we use is less than the
obj_align_mod used for the memory managers.