pdfimage devices - fix %Invocation and add /Info dictionary
I'd thought that we had missed off the %Invocation from the PDF output
of the PDF image devices, but in fact it was only that the pdfocr
devices had it emitted in a surprising place in the file.

That's fixed here just by moving when we write it so that its location
is consistent between the devices.

Moved some definitions common to both pdfimage and pdfocr from the
device's .c files to the pdfimage.h file.

Added emission of a very basic /Info dictionary. To facilitate future
enhancements the number of 'static' (ie predefined) objects has been
turned into a #define and we use that wherever possible instead of the
previous hard-coded magic number. Because of the font objects required
for pdfocr, and the fact that they refer to each other, these still
don't use the #define and will have to be manually adjusted if we ever
change this again.

Acrobat ignores the /Info dictionary, apparently it only reads the XML
/Metadata values (at least in modern versions), so adding the Info
dictionary doesn't help. Creating the XMP metadata is unfortunately
quite complicated, we'd need to steal quite a bit of code from pdfwrite
and at the moment I think its more trouble than its worth. I may change
my mind in future.