"Address" subscript -1 is outside array bounds warning
There is no immediate prospect of resolving the underlying issue: that the
stream code requires the data pointer in the "cursor" to be initialised to the
address one byte *before* the beginning of the actual buffer.

We also do not want to disable the array bounds warning completely, as that
risks missing real mistakes.

So, define a static inline function which does the offending pointer assignment,
and disable the warning locally, for only that function. Then have everywhere
that uses such a buffer setup call that function, rather than do the assignment
itself.

At the moment, it is only disabled for gcc (and compatible) compilers, we can
add others if/as required.