Add support for preserving Link annotations
There is a bunch of PS code that seems to handle some special cases.
I haven't figured it out.  I need to find a sample where the code even
matters.

But this commit will get the basics working, so it is better than nothing.

Don't try to process Dest and StructTreeRoot because they cause
issues.

StructTreeRoot shouldn't even be in a /Link and is super-circular.
See Bug691785.  Need a better solution for dealing with circular stuff.
In theory there could be any kind of garbage in an Annot and right now
I am just trying to deal with all of it.  Maybe I need to whitelist fields
rather than blacklisting them?