Fix handling of circular refs
Added the concept that an indirect_ref can be a label, by adding
is_label flag to them.

This is so I can distinguish the wacky generated labels associated
with high level forms from other cases that are indirect refs that
haven't been resolved because they are circular.

I am not sure this is the best solution. Still need to look at it further.

Basically, for an indirect ref we generate '{N 0 resolveR}' which causes
pdfwrite to generate an object to be referred to. But the object seems
to always be empty, so I am not sure what the point is.

This case only seems to apply specifically to the ref that occurs in
tests_private/pdf/PDF_1.7_FTS/fts_19_1905.pdf