Update tesseract traineddata loader with new path search.
First, we look in TESSDATA_PREFIX (if defined).

If not found there, we look in ROMFS (in tessdata).

If not found there, we look at the configured "tessdata" path
(which defaults to ${datadir}/tessdata). (${datadir} defaults to
${prefix}/share on unix, and ${gsrootdir} on windows.)

If not found there, we look in the current directory.

Update doc/Devices.html (and fix some indexing).