pdfwrite - Terminate pass-through images when we have read enough data
Bug #702964 "EPS file hangs with pdfwrite"

This is only a partial fix for the bug report, the remaining problem
is addressed in a following commit.

The problem here is that the EPS file includes a ProcSet which processes
the JPEG compressed data. It does so using a procedural DataSource for
the image operator. Unfortunately the procedure fails to spot that
readstring returns 0 bytes and identify this as en EOD. Instead it
continues to try and read data, which continuously fails.

While the cause of the hang is in the PostScript program, it does not
normally exhibit because the image operator ceases to read data from the
DataSource when it has read enough samples to satisfy the declared
Width/Height/BPC. An oversight in the pass-through code meant that the
image operator wasn't properly terminating even when enough data had
been supplied.

This commit correctly returns either 0 (require more data) or not-zero
(enough data read) so that the operator terminates.