PostScript interpreter - add a new IdiomSet for PPI 'ColorUtils'
Bug #702964 "EPS file hangs with pdfwrite"

This is not related to the hang, but to the remaining problem with this
file and pdfwrite.

When using pdfwrite with PassThroughJPEGImages the output is incorrect.
The reason is that the PostScript program processes the JPEG compressed
data, turning a 4-component CMYK image into a single component gray
image.

Obviously if we pass through the original CMYK data, but declare the
image a DeviceGray the result is incorrect.

The only way to deal with this is to turn off the feature. There's no
simple way to do this in pdfwrite (where we normally decide to turn it
off, for downsampling etc) without disabling it for a number of other
useful cases. So instead add a new IdiomSet which replaces part of the
PPI ColorUtils ProcSet with a redefined procedure which turns off the
PassThroughJEPGImages feature if the ProcSet forces Gray output.