pdfwrite - reorder code to change evaluation order of image parameters
Bug #702961 "GS produces significantly larger PDFs with same settings as Adobe Distiller"

This is a problem with the interaction of various image parameters. The
relevant options are PassThroughJPEGImages, ColorConversionStrategy and
AutoFilterColorImages.

Because PassThroughJPEGImages is true, we don't set up a compression
filter (we don't want to apply compression to a DCT compressed data
stream). But ColorConversionStrategy is /sRGB, so we then turn off the
PassThroughJPEGImages (we cannot change the colour model and still
pass through the original data, obviously). Ordinarily the compression
chooser would then select a compression based on the result, but
AutoFilterColorImages is false, so the compression chooser is disabled.

The result is that we write uncompressed image data to the file.

By reordering the code so that we check the ColorConversionStrategy
earlier we can turn off PassThroughJPEGImages before selecting the
initial compression which means that this p[articular sequence of
options will then work as expected.