PDF interpreter (with pdfwrite) - another  non-square scaling stroke fix
Bug #702946 "Ghostscript creates PDF originals with black stripes"

This is another case of the problem in bug #702900, The commit to fix
that problem (d5494de2cab28b91ba360d15b8afffef50a3f421) fixed four
places in the code, but missed a fifth place because it is distant from
the other four. The offending procedure is setstrokeforTrpreservation.

This commit adds the required code to that location to go along with the
other four, at the same time:

1) Move the actual code to calculate the width into a procedure
2) Have all 5 places use that procedure
3) Improve the non-square calculation (hypotenuse/2 instead of hypotenuse)
4) Add some 'tolerance' to the equality test.

The tolerance is because both the test file here and in 702900 have a
CTM which is minutely non-square. In this case the c and y scaling
differ by less than 0.0001. The intention of this code is actually to
deal with non-square resolutions, rather than non-square CTM, and it
does so by calculating the hypotenuse of a triangle where the x an y
values are the linewidth in each direction. In fact the hypotenuse/2
seems a closer approximation. But in any event its incorrect, better to
simply use the linewidth if the difference between the two is not very
large. (eventually, with pdfi, we will probably do away with this
horrible kludge entirely)

This causes minor but useful progressions in the test suite files
Bug695841.pdf, clipping_image_problem.pdf and
sumatra/712_-_image_not_displaying.pdf. most notably the latter.

It also produces slightly less bold output for the test file here,
which Acrobat shows as matching precisely with the original.