Bug 702891 and Bug 700593  Fix pdf14 device color procs
The pdf14 color mapping procs clearly had some issues as they are reading uninitialized
memory and using transfer functions that would not make sense with changing
color spaces.   Also, the soft mask should not be going through the effective_transfer.
It has its own transfer function to deal with.  Careful reading of the PDF spec revealed
that the transfer function is only to be used if the color being drawn is opaque.  It
states that this is true if the blend mode is normal, stroke and fill alpha are 1.0, and
no soft mask is present.  Also, an image that has a soft mask does not use the transfer
function.  These changes resulted in a lot of diffs.  I reviewed them and
did some checking of color values to verify things are ok.