Fix Bug702995: Inconsistent auto-rotation of EPS with EPSFitPage.
We need to defer the EPSFitPage scaling, centering and rotation until
after both the the %%BoundingBox and/or the %%HiResBoundingBox have been
processed. Doing one after the other results in slight rounding diffs
depending on the resolution. Change to save the value (prefering the
HiResBoundingBox) and do the scaling/translate/rotate only once when
%%EndComments is processed.