New file - program to assist in creating ZUGFeRD electronic invoices
While documenting the process for creating a ZUGFeRD invoice from a PDF
file and an XML invoice it became clear to me that it was beyond any
reasonable expectation of a user to be able to use it unaided. So this
program assists in the creation of a ZUGFeRD document.

The program requires two command line parameters; -sZUGFeRDProfile=
which contains a fully qualified path to an appropriate (correct colour
space) ICC profile, and -sZUGFeRDXMLFile= which contains a fully
qualified path to the XML invoice file.

Example command line is in the comments, and a usage message if the user
fails to provide either of the required elements. Obviously the user
must also set -dPDFA=3 and -sColorConversionStrategy in order to create
a valid PDF/A-3b file.