Fix indeterminisms within halftoned rendering.
When checking for pdf14 being in an opaque state, we check to see
whether we are either constructing an SMask or are within an
SMask. We need to use a dev spec op for this, as we might be within
a clipper device, and so not actually be directly a pdf14 device.