lcms2: automatically align blocks appropriately on sparc.
The sparc architecture requires pointers within structures to
be at 64bit alignment, even if they are 32bit pointers.

LCMS2 allows for this by having a CMS_PTR_ALIGNMENT symbol
that can be predefined. If it's not predefined, it defaults to
sizeof(void *).

We update the source here so that when building for sparc, it
defaults to 8. This shouldn't affect gs, as it sets the value
via configure/make. I believe our lcms2 repo as used in MuPDF
is autogenerated from this though, and this will help us there.