Fix illegal path operations
When using a type 0 font, or a font with no Widths array, we were
applying TEXT_DO_DRAW which could, in the case of text rendering modes
involving application of stroke, result in an illegal set of path
operations.

Chris Liddell supplied the fix here which is simply to drop the flag
at that point, if we need it then it will be applied by later code.