Fix line widths for stroked text
When stroking test we need to alter the current line width to take
account of the fact that we have merged in the text render matrix, the
text line matrix and the font point size.

We were only partially doing that for the Tj operator and not at all for
the TJ operator.

While an improvement, this does not completely solve the problems noted
with a cluster test, this file:
tests_private/pdf/sumatra/1783_-_CIDToGIDMap_doesn_t_contain_cid.pdf
has too wide a line width. This implies the naive fix here is not
sufficient and we will need to improve it at some point. Its also
likely the code to calculate the new linewidth should be pulled out
into a separate routine.

But the improvement is sufficient its felt worthwhile applying the
code as it stands.