Revise font dir global_glyph_code callback API
The global_glyph_code callback API relies on the name table being avaiable in
the gs_lib_ctx and accessible via the gs_memory_t object.

This is not something that is true, nor can we make it true, for pdfi. Because
pdfi is required to integrate with the Postscript interpreter, we cannot have
the gs_lib_ctx "top_of_system" pointer point to the pdfi context.

So, change the global_glyph_code API so it takes a gs_font pointer, instead of
a gs_memory_t pointer. The Postscript implementation will work exactly the same,
just accessing the gs_memory_t through the gs_font pointer.

But it allows pdfi to grab it's own context through the gs_font "client_data"
pointer.