Fix stream decryption at the file level for indirect objects
There was a misplaced line in pdfi_dereference(). The decrypt_strings
flag was unconditionally reset when it should only have been reset if
the xref entry points to an object in a compressed object stream.

Strings in an ObjStm are not encrypted (the ObjStm is) so we need to
disable decryption when dereferencing such objects. But ordinary
string objects should still be decrypted.