pdfwrite - fix Type 4 Chroma-keyed images with Colour Conversion
Bug #702698 "convert Grayscale error in output"

Images with a /Mask where the Mask is a range of values, and hence a
chroma-keyed image, as opposed to the Mask being an external Image
Mask, and therefore a stencil cannot be preserved as such when we are
colour converting.

There is no reliable way to be certain that the colour of the image
samples after conversion and the converted Mask values will relate in
the same way. Its entirely possible for multiple RGB values to map to
the same gray value for instance, and if that happens to be a masked
value then pixels will be masked which were not originally.

This commit checks the ColorConversionStrategy and if it is not
LeaveColorUnchanged then we further examine the Mask. If the Mask is a
range of values then we consider the colour space of the image.

If ColorConversionStrategy is not LeaveColorunchanged and the image has
a range of values and it is not in the target colour space, then we
cannot preserve it. In this case we fall back to either preserving the
image, and creating a clip path from the chroma key and the image
samples, or (if Version < 1.3) we fall right back to writing each image
sample as a filled rectangle.

This does, of course, result in larger output.