Bug 702671: Make sure X11 device cleans up with closure
The issue is that the pdf14 device will close and reopen the
target device under certain cases and the X11 devices were not
cleaning themselves up sufficiently.   Also added a finalize
method where the  call to XCloseDisplay should actually be made.

The pdf14 device does this close and open dance to ensure that
the page_has_transparency parameter will be set in the device.
It is possible that we could end up in the pdf14 device
push without page_has_transparency if the call is done from
Postscript code.  The PDF interpreter
always sets the page_has_transparency value before doing the
push so this should not be a problem with PDF.

Thanks to Chris Liddell for helping with this.