FIx SEGV with gx_device_unsubclass when child is a clist device.
The 'band_range_list' was a structure of two pointers within the device
(gx_device_clist_writer) so when it was copied, the 'ccl' pointer could
point to the band_range_list structure in the child device. This pointer
would no longer be valid when the child device was freed as the device
unsubclass did.

Detected with 15-01.BIN as it called gx_device_unsubclass for the PCL_mono
subclass device. With -Z@ the band_range_list would be overwritten with
(known, but invalid pointer) data resulting in the SEGV.

Cured by putting the band_range_list into the clist_writer 'data' area.
This area is not GC'ed and since it points into other memory in the
clist writer 'cbuf' area, it is internal to the clist writer.