Fix another SEGV with BGPrint due to device_unsubclass
The printer device needs to have the bg_print structure external to the
device so that when the device is freed the bg_print communication area
shared with the thread doesn't get freed. This is similar to the clist
band_range_list issue.

The SEGV was seen with 15-01.BIN as it unsubclasses the PCL_Mono_Palette
gs_pcl_mono_palette_device device.