Fix bug 702957, 702971. PageList problems with PDF input files.
Thanks to Peter Cherepanov for this fix.

The page skipping was caused by not disabling the PageHabdler in
even or odd page selection modes due to a typo (Pagelist instead of
PageList). The PDF interpreter fed only the odd (or even pages),
and then the PageHandler also skipped every other page.

Also correct log messages and associated operand stack mess-up.