Fix Bug 702034. Missing image to DeviceN devices.
The file from Bug 693300 has a blank image when going to DeviceN devices
such as psdcmyk and tiffsep. The CompatibleOverprint blend mode must be
set before the transparency group is pushed.

Add a .special_op for SupportsDevn to let the PostScript PDF interpreter
detect that the device supports separations.

Make sure the mark fill rect for the knockout case handles
the hybrid case of additive process colors with subtractive spots
and overprint.

And make sure the group that is pushed in gstrans.c for text knockouts
uses compatible overprint if needed.

Ray did the PS parts and Michael did the pdf14dev parts.