Bug 703086 -- Disable trying to preserve Movie annotations
Movie annotations are not implemented.  This fix just disables the attempt to
preserve them, so that the pdfwrite output will be valid.

This file has a /Movie annotation that references a local file that
isn't included in the PDF, so it will never play properly anyway.

The annotation tries to reference a stream in its /Poster entry (for
the image preview of the Movie), and this was not being handled
correctly.