ps2write - fix use of /.HWMargins with ps2write output
Bug #703017 "When print a file created with pdf2ps command with a PDF driver, the image shifts to the upper right as the number of pages increases."

The bug report here is, unfortunately, insufficient to duplicate and
resolve the problem. The missing information was supplied quite
independently by the user 'vyvup' on the #ghostscript IRC channel. See
the #ghostscript logs at around 08:45 on November 9th 2020.

https://ghostscript.com/irclogs/20201109.html

The important missing information is that the device must have the
Ghostscript-specific page device key /.HWMargins set. When this is set
and has non-zero values the offsets are applied cumulatively on every
page by the code in pdfread.ps. This is because we set up the page
'view' before we save the current setup in the ps2write dictionary
under the 'pagesave' key. When we restore back to that at the end of the
page it therefore does not remove the translation applied to account
for the /.HWMargins.

Here we just shift the save so that it occurs before we apply the page
size setup.