PDF interpreter - fix searching for missing Resources in parent object
Bug #703105 "PDF file gives "Unable to determine object number..." and output is missing some images."

As per the bug thread; the PDF file has annotations which are deeply
nested forms. The final form stream draws an Image XObject but the
Form dictionary does not contain a /Resources dictionary so we are unable
to resolve the name.

The form which calls the final form *does* define the missing XObject,
this is pretty clearly illegal but Acrobat copes with it. In fact the
Ghostscript PDF interpreter has code to deal with it too, but there
was a bug in it, it pops an object that was never pushed, resulting in
the code being unable to find the resource.

Fixed very simple here. Also uploaded the simplified file for this bug
and the file for the original bug (700493) to the test repository.