Fix bug 703125: -dNOINTERPOLATE leaves 'true' boolean on the stack.
Thanks to Peter Cherepanov for spotting this. This patch is slightly
different to his in that it makes -dNOINTERPOLATE=false enable interpolation
at the default level (the original patch and the old code would have
disabled interpolation even with 'false'. This version makes the NOINTEPOLATE
and DOINTERPOLATE options operate more symmetrically, however it is sort of
moot since both of these options are intended to be replaced by the better
control on image interpolation provided by -dInterpolateControl=#

Operation tested with fts_17_1712.pdf using the various command line options.