pdfwrite - fix OCR processing
We were not correctly updating the operation and size variables after
we had captured the bitmaps leading the code to error out and drop
content.

Also, if the OCR engine initialisation fails, don't throw an error,
just do nothing. Throwing an error results in us falling back to
rendering the text as bitmaps.