Bug 703146: Fix assert in scan converter.
This is caused by lines changing in x position by more than 1<<31,
and hence the sign bit of an int being confusing. (e.g. the difference
between 0x7fffffff and 0x80000000 is positive, but appears negative
when held in an int). Use an unsigned int instead.

This doesn't affect normal operation as the value is promoted to an
int64_t before being used.