Get Metadata working for pdfwrite
I had to enable "PreserveEPSInfo" and "ParseDSCCommentsForDocInfo" in
order to get this to be automatically triggered in the pdfwrite driver
(see pdf_document_metadata()).

Actually only one of these flags needs to be set, but I have no idea which
one, as it's not documented and the flags aren't really used anywhere for
anything.  In fact the official docs say they don't do anything...
In gs the distillersettings all have this true, so it just works.

This all feels kind of hacky, but it does seem to work....