Change init values of PreserveEPSInfo and ParseDSCCommentsForDocInfo
This is needed for pdfi.

The docs (VectorDevices.htm) say these will defaulted to 'true' and
the gs/PostScript code sets them to true, but they were initialized to
false in the device itself.

The only place these variables are actually used is in
pdf_document_metadata() where they are tested to decide whether to
emit the Metadata.

This change makes no difference for gs and gs/pdf, but it will cause
gpcl to start including Metadata in its pdfwrite output.