A little more tidying up after the refactor
Use the pdfi_stream_offset() accessor instead of poking in the
pdf_stream object directly in two places.

Dispense with a number of casts.

Remove a prototype for a function that no longer exists.

Remove some #if 0'ed code.

Add a debug warning to pdfi_obj_dict_to_stream() to try and spot usage
of an incompatible dictionary (ie a potentially cached one).