oss-fuzz 27011: Improve handling of error during a "save".
Previously, in the event of an error during a Postscript save operation, we
could end up leaving a half formed save object in existence. If the error
happened during the gs_gsave_for_save() function, that could end up leaving
the graphics state cleanup "out of sync" with the save state cleanup at the
final virtual memory cleanup as we shut down.

To resolve that, destroy the partially formed save state at the point of the
error. This also means tweaking dorestore() to handle unrolling a save state
without the graphics state reference in it.