Bug 703164: Endian issues with CMM
The interface code to the CMM was corrected to indicate when a
endian swap was needed on the data.  This should only occur
in the case when we are dealing with transparency buffers
during the put image blending operation that may include
a color conversion.  The final blend bakes the data as BE
so if we are on a LE machine, the CMM will need to know to
swap the bytes (assuming the pdf14 device is using 16bit buffers).

The code was rewritten to make it clear that this setting is no
BE vs LE but simply an endian swap.  That was a source of confusion.

Revealed in this testing was the lack of some proper error
reporting during buffer conversions, which were fixed.