Fix reference counting of stream 'parent_obj', simplify code.
This is the code to deal with a stream which uses a named resource which
is not defined in its own Resources dictionary, but is defined in the
Resources dictionary of a stream which is being executed, and the
current stream is executing as part of that stream's content.

If we count up the parent object when we store it in the child stream's
structure (as would seem right), then we can end up with a circular
reference. The parent stream's Resources dictionary contains a
reference to the child stream and the child stream holds a reference to
the parent. The result is that each one has a minimum reference count of
1.

The simplest way to deal with this is to note that we only care about
(and potentially use) the parent when we are actually executing the
content, so we now set and reset the parent object during the course
of executing the stream, the parent_obj should only ever be non-NULL
when the stream is being executed.

This has the side benefit of simplifying the code quite a bit, we no
longer need to worry about the parent_obj in dereferencing. I took the
opportunity to simplify pdfi_interpret_content_stream a little too.

Added a bunch more comments to try and explain this ugliness.