pdfwrite - fix 2 memory leaks with metadata
Discovered with pdfi. There is a temporary buffer which wasn't being
freed.

A more complex problem involves the Metadata value stored in the
Catalog dictionary. A long comment describes why this is not freed as
an 'otherResource', but experience contradicts the comment, the
Metadata value is not freed along with the Catalog object.

I've altered this here to solve the memory leak but this may need more
investigation.