For text mode 3 we were inadvertently setting the text 'operation' flags to a particular value instead of OR'ing in the flags we wanted.
The result of this was that we had no TEXT_DO_* flags which meant the
operation was rejected by the graphics library as illegal. This meant
we were dropping Tr 3 text which is invisible and therefore not
apparently incorrect. However the output from pdfwrite was missing
the text.

I've also put back the error flag when setting an invalid text matrix
in Tm because it doesn't get set in show, as I thought it did.