Ensure the nulldevice has a proper set of device methods
gs_make_null_device uses the gs_null_device structure to create an
instance of the nulldevice, but it did not call gx_device_fill_in_procs
to set the device methods which are left as NULL in the declaration.

Since all device methods (except fill_rectangle) are expected to be
real functions, not NULL, many places in the code do not check that
the device method is not NULL before calling it. This could (and in
pdfi did) lead to us attempting to execute 0x00 causing seg faults.