Fix memory leak with copied type 1 fonts.
If we copy a type 1 font (for the benefit of the pdfwrite family of
devices) then we make a copy of the Subrs and GSubrs as well, but the
routine to free the copied font wasn't freeing those copies.

This isn't normally obvious because these are garbage collected, found
while checking pdfi.