Use the correct type for initializing a structure
Use the length of the charstring data, rather than the (possibly larger) size
of the available buffer, when decoding charstring data.

The above solve a large number of segfaults when calling pdfi from Ghostscript.

Also quiet a couple of compiler warnings noticed in passing and use the
correct offsets when building the fontmap buffer for parsing.