Fix bug 703198: SEGV's with psdcmykog device and --saved-pages-test
This was a reference of a missing icc_table caused by the band count not
being adjusted during saved-page playback as it was during writing. The
orig_spec_op in the gx_device_printer structure was being set to the
gdev_prn_spec_op, but this did not call the target device's spec_op, so
it did not get the result from gxdso_adjust_bandheight that was needed.

Caused by commit 8df410c26.

Also clean up a left over extern in gxclrect.c (not related to the fix).