pdfwrite - fix memory leak with high level forms
As usual this does not exhibit with Ghostscript, because of garbage
collection, and was observed with pdfi.

We transferred ownership of the form object to the local named
resources dictionary, and removed the resource from the XObject chain
of stored resources, but we didn't free the actual resource structure.

Rejigged the code order slightly because we need the resource for a
few purposes before we transfer ownership and free it. Also fixed a
typo and a misleading string (only used for debug purposes in each case)