Change the handling of references to free objects
No bug report as yet, mail from Anja Brakensiek of OpenText on Dec 1st
2020.

Previously when a request was made to dereference an object, and the
object was marked as free in the xref table we would return a null
object. Its clear for the test file that Acrobat attempts to use the
'next object number' of the free entry as an offset which is crazy but
because this is the first object in the file actually works.

We now mimic this, we try to use the 'object number' as an offset. If
we fail to read an object, or the object we read has the wrong object
number, we still return a null object.