Bug 694692: Handle custom left-sidebearing values.
Ghostscript was ending up ignoring LSB values read from Metrics dictionaries.

This stemmed from some confusion about Freetype's incremental interface
API get_metrics call. We had taken the references to "overriding" the metrics
to mean it allowed for custom metrics to be passed into (and back out of)
Freetype. That is not the case: it is intended to allow integrations supporting
"incomplete" (primarily TrueType) fonts, such as PCL/XL embedded TTFs, that are
permitted to omit certain tables - specifically, in this case, the hmtx and vmtx
tables.

So, this drops that approach, and applies custom LSB values in our code.

Coincidentally, this also makes gs somewhat immune from a pending change in
Freetype related to that get_metrics call.