Take out code to initialize some pdfwrite device params
This code is no longer needed because the values of these parameters
are now 'true' in the pdfwrite device.  Which is how it should have
been all along.

I am leaving the code in here but disabled for now, since it is a
good place to handle similar device initializations if necessary in
the future.