Bug 703214 - treat some kinds of free objects as not free
The file in Bug 703214 has been edited incrementally a number of times.
The bug is actually the creation of the initial file, which declares
its xref table as having a section 7 entries long starting at object
1.

In fact, as is normal, the xref table starts at object 0 with the free
entry at the start of the linked list. Because of the incorrect
subsection header though, a PDF consumer should read this as being
object 1, not object 0.

Obviously *all* the objects are then incorrect.

The incremental updates then redefined very object *except* object 1.
The result of this is that these objects are all correctly defined. The
file then attempts to render object 1, which is free. Acrobat can deal
with this, which must mean it is ignoring the 'free' status of the
object and reading the 'next free object' value as if it were the offset
of the object.

Because the next free object is 0, and object 1 is the first object in
the PDF file, and the preceding line is a comment, this actually works
though I'm pretty certain its only by sheer luck, it really should not
work at all.

Unfortunately, because of the way the the PDF interpreter reads PDF
files and stores xref entires, we cannot easily detect an attempt to
use a 'free' object and try to treat it as non-free. We have to do this
upfront when we read the xref table. Because this is such a fluke and I
believe not likely to be repeatable, I've tried to keep the condition
for ignoring the free flag as tight as possible; The generation number
must be exactly 65535 (the canonical value for the head of the free
list), and the object number must be not zero. In this case only we
will store the 'next free object number' as an offset.

If the object is truly free then either it is unused (in which case we
will not use the xref entry) or it should have been superseded by a
later incremental update, in which case we still won't use this entry.
The only way we could try to use this entry is if the PDF file attempts
to use a free entry.

This might change some handling of other broken files.